GH-165: Simplified and renamed the editor-utilities, based on feedback from akosyakov
